server/requirements.txt
beautifulsoup4==4.13.3
chromadb==0.6.3
duckduckgo_search==7.3.2
fastapi==0.115.8
groq==0.18.0
langchain==0.3.18
python-dotenv==1.0.1
Requests==2.32.3
uvicorn==0.34.0
